moving free lists into a separate file.
